+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes on the registration page.
 
 - Add request timeout notification request.
 - Fix js IE issue (no const, use var in twig template)
+- Add authenticateUrl to authentication page 
 
 ## 2.1.1
 Changes on the registration page.